src/python/lambda/system_pi/requirements.txt
requests
mysql-connector
datetime_truncate
numpy
pandas
boto3
